refactor: Pull CSV to object methods into DataFactory class file and refactor w/ parent class
